move most file into a lib directory
