+++ lib/cli-view.js
